Implement comunication between server and client, the client is updated based don server data (partially for now...)
I added the Access-Control-Allow-Origin header to ech route of the server (since is gonna run only locally) in order to allow the comunication between the flask application and the react application.

Now the list of results is fetched as soon as the page loads and it gets displayed in a correct way.

The warnings, relative to a single result, get fetched asynchronously when the user expands the card of the result (they are not visualized yet, just shown on the JS console...).

Signed-off-by: Sebastiano Mariani <mariani.sebastiano@gmail.com>