Finish POC, it misses only the info and the context splitting
The POC of the entire application is finished (it misses the two things in the title). The code sucks and the app is slow (no pagination etc...). The entire application needs a refactory but I want to work on something else now so I will leave those improvements to someone else or I will do them myself in a remote future...

Signed-off-by: Sebastiano Mariani <mariani.sebastiano@gmail.com>